+++ Passepartout-iOS/Global/SwiftGen+Assets.swift
@@ -283,6 +283,7 @@ internal enum Asset {
   internal enum Providers {
     internal static let mullvad = ImageAsset(name: "mullvad")
     internal static let pia = ImageAsset(name: "pia")
+    internal static let protonvpn = ImageAsset(name: "protonvpn")
     internal static let tunnelbear = ImageAsset(name: "tunnelbear")
     internal static let windscribe = ImageAsset(name: "windscribe")
   }