+++ Passepartout/Sources/Services/InfrastructureFactory.swift
@@ -59,6 +59,7 @@ public class InfrastructureFactory {
     public let allNames: [Infrastructure.Name] = [
         .mullvad,
         .pia,
+        .protonVPN,
         .tunnelBear,
         .windscribe
     ]